scx_lavd: Reorganize the latency/performance criticality code.
As the code related to the latency & performance criticality calculation
is large enough, let's separate it out from the main.bpf.c and create
lat_cri.bpf.c. This is a simple reorganization of the code without any
functional changes.
